working X3100
git-svn-id: http://subversion.assembla.com/svn/fakesmc/HWSensors4@530 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ FakeSMC/FakeSMCKey.h
@@ -22,7 +22,7 @@ inline void copySymbol(const char *from, char* to)
 {
 	bzero(to, 5);
 	
-	UInt64 len = strlen(from);
+	UInt32 len = (UInt32)strlen(from);
 	
 	bcopy(from, to, len > 4 ? 4 : len);
 }
